Add new file for mimir's EquiJoinPlanner.
